# Conflicts:
#	app/controllers/api/lti_applications_instances_controller.rb
#	client/js/actions/instances.js
#	client/js/admin_routes.jsx
#	client/js/components/admin/common/heading.jsx
#	client/js/components/admin/instances/_instances.jsx
#	client/js/components/admin/instances/instance.jsx
#	client/js/components/admin/instances/instance_header.jsx
#	client/js/components/admin/lti_applications/application_row.jsx
#	client/js/components/admin/lti_applications/index.jsx
#	client/js/components/layout/index.jsx

+++ lib/middleware/oauth_state_middleware.rb
@@ -5,14 +5,23 @@ class OauthStateMiddleware
 
   def call(env)
     request = Rack::Request.new(env)
-    if request.params["state"]
+    if request.params["state"] && request.params["code"]
       if oauth_state = OauthState.find_by(state: request.params["state"])
-        env["oauth.state"] = JSON.parse(oauth_state.payload) || {}
-        application_instance = ApplicationInstance.find_by(lti_key: env["oauth.state"]["oauth_consumer_key"])
+        # Restore the param from before the OAuth dance
+        state_params = JSON.parse(oauth_state.payload) || {}
+        state_params.each do |key, value|
+          request.update_param(key, value)
+        end
+        application_instance = ApplicationInstance.find_by(lti_key: state_params["oauth_consumer_key"])
         env["canvas.url"] = application_instance.lti_consumer_uri
         oauth_state.destroy
+      else
+        raise OauthStateMiddlewareException, "Invalid state in OAuth callback"
       end
     end
     @app.call(env)
   end
+end
+
+class OauthStateMiddlewareException < RuntimeError
 end
